Fix build

+++ source/game_sa/Tasks/TaskTypes/TaskComplexShuffleSeats.cpp
@@ -47,7 +47,7 @@ CTask* CTaskComplexShuffleSeats::CreateSubTask(eTaskType taskType, CPed* ped) {
     case TASK_SIMPLE_CAR_SET_PED_IN_AS_DRIVER:
         return new CTaskSimpleCarSetPedInAsDriver{m_veh, m_utilityLineUp};
     case TASK_SIMPLE_CAR_SHUFFLE:
-        return new CTaskSimpleCarShuffle{ m_veh, m_startDoor, m_utilityLineUp };
+        return new CTaskSimpleCarShuffle{ m_veh, (int32)m_startDoor, m_utilityLineUp };
     case TASK_SIMPLE_CAR_SET_PED_IN_AS_PASSENGER:
         return new CTaskSimpleCarSetPedInAsPassenger{ m_veh, m_targetDoor, false, m_utilityLineUp };
     case TASK_SIMPLE_CAR_SET_PED_OUT:
